feat: 알림 등록 기능, 테스트 코드 구현

+++ src/main/java/com/redhorse/deokhugam/domain/dashboard/dto/popularreview/PopularReviewDto.java
@@ -13,7 +13,7 @@ public record PopularReviewDto(
         String bookThumbnailUrl,
         UUID userId,
         String userNickname,
-        int reviewCount,
+        long reviewCount,
         double reviewRating,
         PeriodType period,
         Instant createdAt,
